integrationtests/requirements.txt
oslo.config>=1.2.0
python-keystoneclient>=0.6.0
testresources>=0.2.4
testtools>=0.9.34
python-glanceclient>=0.9.0
simplejson>=2.0.9
jsonpatch>=1.1
jsonpath-rw>=1.2.0,<2.0
jsonrpclib
jsonschema>=2.0.0,<3.0.0
python-muranoclient
murano-metadataclient
